rogue_player.h
