+++ src/cc/rogue/rogue.h
@@ -357,20 +357,7 @@ union _bits256 { uint8_t bytes[32]; uint16_t ushorts[16]; uint32_t uints[8]; uin
 typedef union _bits256 bits256;
 #endif
 
-
-#ifndef ROGUE_DECLARED_PACK
-struct rogue_packitem
-{
-    int32_t type,launch,count,which,hplus,dplus,arm,flags,group;
-    char damage[8],hurldmg[8];
-};
-struct rogue_player
-{
-    int32_t gold,hitpoints,strength,level,experience,packsize,dungeonlevel,pad;
-    struct rogue_packitem roguepack[MAXPACK];
-};
-#define ROGUE_DECLARED_PACK
-#endif
+#include "rogue_player.h" // interface to rpc
 
 struct rogue_state
 {
